match directory name to package name - needed for pypi package
